% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{execute}
\alias{execute}
\title{Execute OHDSI DOSETTE study}
\usage{
execute(
  connectionDetails,
  cdmDatabaseSchema,
  vocabularyDatabaseSchema = cdmDatabaseSchema,
  oracleTempSchema = NULL,
  cohortDatabaseSchema,
  outputFolder,
  tablePrefix = "dosette",
  databaseId = "Unknown",
  databaseName = "Unknown",
  databaseDescription = "Unknown",
  minCellCount = 5,
  createExposureCohorts = TRUE,
  createOutcomeCohorts = TRUE,
  createNegativeControlCohorts = TRUE,
  generateCensorAnalysisObjects = TRUE,
  pullAnalysisCohortsFromServer = TRUE,
  runCensorWeightAnalysis = TRUE,
  exportToCsv = TRUE,
  maxCores = 4
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created
using the
\code{\link[DatabaseConnector]{createConnectionDetails}}
function in the DatabaseConnector package.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM
format resides. Note that for SQL Server, this should
include both the database and schema name, for example
'cdm_data.dbo'.}

\item{vocabularyDatabaseSchema}{Schema name where your vocabulary tables in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{oracleTempSchema}{Should be used in Oracle to specify a schema where the
user has write priviliges for storing temporary tables.}

\item{cohortDatabaseSchema}{Schema name where intermediate data can be stored. You
will need to have write priviliges in this schema. Note
that for SQL Server, this should include both the
database and schema name, for example 'cdm_data.dbo'.}

\item{outputFolder}{Name of local folder to place results; make sure to use
forward slashes (/). Do not use a folder on a network
drive since this greatly impacts performance.}

\item{tablePrefix}{A prefix to be used for all table names created for this
study.}

\item{databaseId}{A short string for identifying the database (e.g.
'Synpuf').}

\item{databaseName}{The full name of the database (e.g. 'Medicare Claims
Synthetic Public Use Files (SynPUFs)').}

\item{databaseDescription}{A short description (several sentences) of the database.}

\item{minCellCount}{The minimum cell count for fields contains person counts
or fractions when exporting to CSV.}

\item{createExposureCohorts}{Create the tables with the exposure cohorts?}

\item{createOutcomeCohorts}{Create the tables with the outcome cohorts?}

\item{createNegativeControlCohorts}{Create the tables with the negative control outcome cohorts?}

\item{generateCensorAnalysisObjects}{Manipulate cohort data into data frames for censor weight analysis?}

\item{pullAnalysisCohortsFromServer}{Join exposure-outcome cohorts and download from server?}

\item{runCensorWeightAnalysis}{Run the clone censor weight analysis for TTE?}

\item{exportToCsv}{Export all results to CSV files?}

\item{maxCores}{How many parallel cores should be used? If more cores
are made available this can speed up the analyses.}
}
\description{
Execute OHDSI DOSETTE study
}
\details{
This function executes the OHDSI DOSETTE study.
}
